barcode added to Cloud

+++ app/controllers/products_controller.rb
@@ -37,16 +37,15 @@ class ProductsController < ApplicationController
 
       barcode1 =  Barby::Code128B.new(barcodeName)
 
-      File.open( "app/assets/images#{barcodeName}.png", 'w'){|f|
+      File.open( "app/assets/images/#{barcodeName}.png", 'w'){|f|
         f.write barcode1.to_png(:height => 20, :width => 20, :margin => 5)
       }
 
-      @product.avatar = "#{barcodeName}.png"
-
-      Cloudinary::Uploader.upload("app/assets/images/#{barcodeName}.png", :public_id => "#{barcodeName}")
+      cloud_barcode_name = Cloudinary::Uploader.upload("app/assets/images/#{barcodeName}.png")
+      @product.avatar = "#{cloud_barcode_name["public_id"]}.png"
 
       if @product.save
-        File.delete("app/assets/images/#{barcodeName}.png")
+        File.delete("app/assets/images/#{barcodeName}.png") if File.exist?("app/assets/images/#{barcodeName}.png")
 
         if params[:purchase_order].present?
           purchase_order = PurchaseOrder.new
